Fix jpg optimization
I used the same way of png copy file. With the old way, I don't know why, the mime type was changed to "application/octet-stream" and it was impossible to view the image again.

I also added a -perfect option which check that -copy and -optimize job is well done.